requirements.txt
ctransformers==0.2.5
sentence-transformers>=2.6.0
pinecone
langchain==0.0.225
flask
pypdf
python-dotenv
cryptography>=3.1
-e .